flask-backend/requirements.txt
torch
torchvision
pillow
tqdm
pandas
flask
flask_cors
